(docs) rename files
