+++ website/src/pages/ask-me.tsx
@@ -2,8 +2,8 @@ import useBaseUrl from '@docusaurus/useBaseUrl'
 import SearchIcon from '@site/static/img/search-icon.svg'
 import Layout from '@theme/Layout'
 import React, { useState } from 'react'
-import { ChatResponse } from '../components/booster-chat/ChatResponse'
-import { CoolTitle } from '../components/booster-chat/CoolTitle'
+import { ChatResponse } from '../components/BoosterChat/ChatResponse'
+import { CoolTitle } from '../components/BoosterChat/CoolTitle'
 import { ChatSerivce } from '../services/chat-service'
 
 const NO_RESPONSE = 'Sorry, I don`t know how to help with that.'
